+++ wod_predictor/feature_engineering_parts/athlete_info_fe.py
@@ -1,6 +1,6 @@
 import pandas as pd
 from .base import BaseFEPipelineObject
-from .helpers import fill_missing_values, convert_units
+from .utils.helpers import fill_missing_values, convert_units
 from ..constants import Constants as c
 from datetime import datetime
